fix(cohere): adjust event emission logic
