+++ packages/opentelemetry-instrumentation-cohere/opentelemetry/instrumentation/cohere/span_utils.py
@@ -1,0 +1,168 @@
+from opentelemetry.instrumentation.cohere.utils import (
+    dont_throw,
+    should_send_prompts,
+)
+from opentelemetry.semconv._incubating.attributes.gen_ai_attributes import (
+    GEN_AI_RESPONSE_ID,
+)
+from opentelemetry.semconv_ai import (
+    LLMRequestTypeValues,
+    SpanAttributes,
+)
+from opentelemetry.trace.status import Status, StatusCode
+
+
+def _set_span_attribute(span, name, value):
+    if value is not None:
+        if value != "":
+            span.set_attribute(name, value)
+    return
+
+
+@dont_throw
+def set_input_attributes(span, llm_request_type, kwargs):
+    if not span.is_recording():
+        return
+
+    if should_send_prompts():
+        if llm_request_type == LLMRequestTypeValues.COMPLETION:
+            _set_span_attribute(span, f"{SpanAttributes.LLM_PROMPTS}.0.role", "user")
+            _set_span_attribute(
+                span, f"{SpanAttributes.LLM_PROMPTS}.0.content", kwargs.get("prompt")
+            )
+        elif llm_request_type == LLMRequestTypeValues.CHAT:
+            _set_span_attribute(span, f"{SpanAttributes.LLM_PROMPTS}.0.role", "user")
+            _set_span_attribute(
+                span, f"{SpanAttributes.LLM_PROMPTS}.0.content", kwargs.get("message")
+            )
+        elif llm_request_type == LLMRequestTypeValues.RERANK:
+            for index, document in enumerate(kwargs.get("documents")):
+                _set_span_attribute(
+                    span, f"{SpanAttributes.LLM_PROMPTS}.{index}.role", "system"
+                )
+                _set_span_attribute(
+                    span, f"{SpanAttributes.LLM_PROMPTS}.{index}.content", document
+                )
+
+            _set_span_attribute(
+                span,
+                f"{SpanAttributes.LLM_PROMPTS}.{len(kwargs.get('documents'))}.role",
+                "user",
+            )
+            _set_span_attribute(
+                span,
+                f"{SpanAttributes.LLM_PROMPTS}.{len(kwargs.get('documents'))}.content",
+                kwargs.get("query"),
+            )
+
+
+@dont_throw
+def set_response_attributes(span, llm_request_type, response):
+    if not span.is_recording():
+        return
+    if should_send_prompts():
+        if llm_request_type == LLMRequestTypeValues.CHAT:
+            _set_span_chat_response(span, response)
+        elif llm_request_type == LLMRequestTypeValues.COMPLETION:
+            _set_span_generations_response(span, response)
+        elif llm_request_type == LLMRequestTypeValues.RERANK:
+            _set_span_rerank_response(span, response)
+
+    span.set_status(Status(StatusCode.OK))
+
+
+def set_span_request_attributes(span, kwargs):
+    if not span.is_recording():
+        return
+
+    _set_span_attribute(span, SpanAttributes.LLM_REQUEST_MODEL, kwargs.get("model"))
+    _set_span_attribute(
+        span, SpanAttributes.LLM_REQUEST_MAX_TOKENS, kwargs.get("max_tokens_to_sample")
+    )
+    _set_span_attribute(
+        span, SpanAttributes.LLM_REQUEST_TEMPERATURE, kwargs.get("temperature")
+    )
+    _set_span_attribute(span, SpanAttributes.LLM_REQUEST_TOP_P, kwargs.get("top_p"))
+    _set_span_attribute(
+        span, SpanAttributes.LLM_FREQUENCY_PENALTY, kwargs.get("frequency_penalty")
+    )
+    _set_span_attribute(
+        span, SpanAttributes.LLM_PRESENCE_PENALTY, kwargs.get("presence_penalty")
+    )
+
+
+def _set_span_chat_response(span, response):
+    index = 0
+    prefix = f"{SpanAttributes.LLM_COMPLETIONS}.{index}"
+    _set_span_attribute(span, f"{prefix}.content", response.text)
+    _set_span_attribute(span, GEN_AI_RESPONSE_ID, response.response_id)
+
+    # Cohere v4
+    if hasattr(response, "token_count"):
+        _set_span_attribute(
+            span,
+            SpanAttributes.LLM_USAGE_TOTAL_TOKENS,
+            response.token_count.get("total_tokens"),
+        )
+        _set_span_attribute(
+            span,
+            SpanAttributes.LLM_USAGE_COMPLETION_TOKENS,
+            response.token_count.get("response_tokens"),
+        )
+        _set_span_attribute(
+            span,
+            SpanAttributes.LLM_USAGE_PROMPT_TOKENS,
+            response.token_count.get("prompt_tokens"),
+        )
+
+    # Cohere v5
+    if hasattr(response, "meta") and hasattr(response.meta, "billed_units"):
+        input_tokens = response.meta.billed_units.input_tokens
+        output_tokens = response.meta.billed_units.output_tokens
+
+        _set_span_attribute(
+            span,
+            SpanAttributes.LLM_USAGE_TOTAL_TOKENS,
+            input_tokens + output_tokens,
+        )
+        _set_span_attribute(
+            span,
+            SpanAttributes.LLM_USAGE_COMPLETION_TOKENS,
+            output_tokens,
+        )
+        _set_span_attribute(
+            span,
+            SpanAttributes.LLM_USAGE_PROMPT_TOKENS,
+            input_tokens,
+        )
+
+
+def _set_span_generations_response(span, response):
+    _set_span_attribute(span, GEN_AI_RESPONSE_ID, response.id)
+    if hasattr(response, "generations"):
+        generations = response.generations  # Cohere v5
+    else:
+        generations = response  # Cohere v4
+
+    for index, generation in enumerate(generations):
+        prefix = f"{SpanAttributes.LLM_COMPLETIONS}.{index}"
+        _set_span_attribute(span, f"{prefix}.content", generation.text)
+        _set_span_attribute(span, f"gen_ai.response.{index}.id", generation.id)
+
+
+def _set_span_rerank_response(span, response):
+    _set_span_attribute(span, GEN_AI_RESPONSE_ID, response.id)
+    for idx, doc in enumerate(response.results):
+        prefix = f"{SpanAttributes.LLM_COMPLETIONS}.{idx}"
+        _set_span_attribute(span, f"{prefix}.role", "assistant")
+        content = f"Doc {doc.index}, Score: {doc.relevance_score}"
+        if doc.document:
+            if hasattr(doc.document, "text"):
+                content += f"\n{doc.document.text}"
+            else:
+                content += f"\n{doc.document.get('text')}"
+        _set_span_attribute(
+            span,
+            f"{prefix}.content",
+            content,
+        )
