+++ packages/opentelemetry-instrumentation-cohere/opentelemetry/instrumentation/cohere/utils.py
@@ -1,13 +1,11 @@
 import logging
+import os
 import traceback
-from os import environ
 
-from opentelemetry._events import EventLogger
+from opentelemetry import context as context_api
 from opentelemetry.instrumentation.cohere.config import Config
 
-OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT = (
-    "OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT"
-)
+TRACELOOP_TRACE_CONTENT = "TRACELOOP_TRACE_CONTENT"
 
 
 def dont_throw(func):
@@ -35,12 +33,10 @@ def dont_throw(func):
     return wrapper
 
 
-def is_content_enabled() -> bool:
-    capture_content = environ.get(
-        OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT, "false"
-    )
-
-    return capture_content.lower() == "true"
+def should_send_prompts():
+    return (
+        os.getenv(TRACELOOP_TRACE_CONTENT) or "true"
+    ).lower() == "true" or context_api.get_value("override_enable_content_tracing")
 
 
 def should_emit_events() -> bool:
@@ -48,6 +44,4 @@ def should_emit_events() -> bool:
     Checks if the instrumentation isn't using the legacy attributes
     and if the event logger is not None.
     """
-    return not Config.use_legacy_attributes and isinstance(
-        Config.event_logger, EventLogger
-    )
+    return not Config.use_legacy_attributes
