+++ packages/opentelemetry-instrumentation-cohere/tests/conftest.py
@@ -5,9 +5,7 @@ import os
 import cohere
 import pytest
 from opentelemetry.instrumentation.cohere import CohereInstrumentor
-from opentelemetry.instrumentation.cohere.utils import (
-    OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT,
-)
+from opentelemetry.instrumentation.cohere.utils import TRACELOOP_TRACE_CONTENT
 from opentelemetry.sdk._events import EventLoggerProvider
 from opentelemetry.sdk._logs import LoggerProvider
 from opentelemetry.sdk._logs.export import (
@@ -68,7 +66,7 @@ def instrument_legacy(tracer_provider):
 
 @pytest.fixture(scope="function")
 def instrument_with_content(tracer_provider, event_logger_provider):
-    os.environ.update({OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT: "True"})
+    os.environ.update({TRACELOOP_TRACE_CONTENT: "True"})
 
     instrumentor = CohereInstrumentor(use_legacy_attributes=False)
     instrumentor.instrument(
@@ -78,13 +76,13 @@ def instrument_with_content(tracer_provider, event_logger_provider):
 
     yield instrumentor
 
-    os.environ.pop(OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT, None)
+    os.environ.pop(TRACELOOP_TRACE_CONTENT, None)
     instrumentor.uninstrument()
 
 
 @pytest.fixture(scope="function")
 def instrument_with_no_content(tracer_provider, event_logger_provider):
-    os.environ.update({OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT: "False"})
+    os.environ.update({TRACELOOP_TRACE_CONTENT: "False"})
 
     instrumentor = CohereInstrumentor(use_legacy_attributes=False)
     instrumentor.instrument(
@@ -94,7 +92,7 @@ def instrument_with_no_content(tracer_provider, event_logger_provider):
 
     yield instrumentor
 
-    os.environ.pop(OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT, None)
+    os.environ.pop(TRACELOOP_TRACE_CONTENT, None)
     instrumentor.uninstrument()
